T222 Move plot functions to separate file
